made nav link and contact form as responsive
