+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import MainContentLayout from "./components/content/MainContentLayout";
-import LeftNav from "./components/LeftNav";
+import MenuLayout from "./components/menu/MenuLayout";
 import ContactContentLayout from "./components/content/ContactContentLayout";
 
 function App() {
@@ -12,10 +12,10 @@ function App() {
     return <MainContentLayout />;
   };
   return (
-    <div className="min-h-screen w-full flex flex-col bg-white dark:bg-black text-black dark:text-white">
-      <div className="flex flex-1 flex-row">
-        <LeftNav onClick={setLinkClicked} linkClicked={linkClicked} />
-        <div className="bg-gradient-to-b from-orange-500 via-red-600 to-blue-800 w-1"></div>
+    <div className="min-h-screen w-full flex md:flex-col flex-row bg-white dark:bg-black text-black dark:text-white">
+      <div className="flex flex-1 md:flex-row flex-col">
+        <MenuLayout onClick={setLinkClicked} linkClicked={linkClicked} />
+        <div className="hidden md:block bg-gradient-to-b from-orange-500 via-red-600 to-blue-800 w-1"></div>
         <div className="grow flex flex-col justify-center items-center">
           {renderContent()}
         </div>
