+++ src/components/content/MainContentLayout.tsx
@@ -1,9 +1,9 @@
 const MainContentLayout = () => {
   return (
     <>
-      <p className="text-4xl font-bold">Hey Folk's</p>
-      <p className="text-6xl mt-4">This is Ajith VM</p>
-      <p className="mt-2 text-gray-600 dark:text-gray-300">
+      <p className="md:text-4xl sm:text-xl text-sm font-bold">Hey Folk's</p>
+      <p className="md:text-6xl sm:text-4xl text-xl mt-4">This is Ajith VM</p>
+      <p className="md:text-base sm:text-sm text-xs mt-2 text-gray-600 dark:text-gray-300">
         Currently working as software engineer @
         <a href="https://www.epam.com" className="text-colorCode">
           EPAM Systems
